- separated filing-interfaces.ts into new files - deleted unnecessary IncorporationFilingBodyIF - added some types to FilingDataIF - updated buildIaFilin()g to use new IncorporationFilingIF and handle updated share structure/classes - updated parseIncorpApp() to handle updated share structure/classes - updated formatEmptyFiling() to handle updated share structure/classes
