+++ src/interfaces/filing-interfaces/filingData-interface.ts
@@ -1,7 +1,10 @@
+import { FilingCodes, EntityTypes } from '@/enums'
+
+/** Filing data object passed to the SBC Fee Summary. */
 export interface FilingDataIF {
-  filingDescription?: string;
-  filingTypeCode: string;
-  entityType: string;
+  filingDescription?: string
+  filingTypeCode: FilingCodes
+  entityType: EntityTypes
   waiveFees?: boolean;
   futureEffective?: boolean
 }
