+++ src/interfaces/filing-interfaces/index.ts
@@ -1,2 +1,2 @@
-export * from './filing-interface'
 export * from './filingData-interface'
+export * from './incorporationFiling-interface'
