+++ src/interfaces/store-interfaces/state-interfaces/businessInformation-interface.ts
@@ -1,0 +1,4 @@
+export interface BusinessInformationIF {
+  legalType: string
+  identifier: string
+}
