+++ src/interfaces/store-interfaces/state-interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from './account-information-interface'
+export * from './businessInformation-interface'
 export * from './name-request-interface'
 export * from './tombstone-interface'
