+++ src/mixins/filing-template-mixin.ts
@@ -3,8 +3,7 @@ import { Component, Vue } from 'vue-property-decorator'
 import { State, Getter, Action } from 'vuex-class'
 
 // Interfaces
-import { ActionBindingIF, StateModelIF, IncorporationFilingIF, IncorporationFilingBodyIF, OrgPersonIF, ShareClassIF
-} from '@/interfaces'
+import { ActionBindingIF, StateModelIF, IncorporationFilingIF, OrgPersonIF, ShareClassIF } from '@/interfaces'
 
 // Constants
 import { INCORPORATION_APPLICATION } from '@/constants'
@@ -48,48 +47,51 @@ export default class FilingTemplateMixin extends Vue {
   @Action setFilingDate!: ActionBindingIF
   @Action setIncorporationAgreementStepData!: ActionBindingIF
 
-  /** Constructs an Incorp App filing body from store data. Used when saving a filing. */
+  /**
+   * Builds an Incorporation Application filing body from store data. Used when saving a filing.
+   * @returns the IA filing body to save
+   */
   buildIaFiling (): IncorporationFilingIF {
     // Build filing.
     const filing: IncorporationFilingIF = {
-      filing: {
-        header: {
-          name: INCORPORATION_APPLICATION,
-          certifiedBy: this.stateModel.certifyState.certifiedBy,
-          date: this.getFilingDate || this.getCurrentDate,
-          folioNumber: this.getFolioNumber,
-          isFutureEffective: this.isFutureEffective
+      header: {
+        name: INCORPORATION_APPLICATION,
+        certifiedBy: this.stateModel.certifyState.certifiedBy,
+        date: this.getFilingDate || this.getCurrentDate,
+        folioNumber: this.getFolioNumber,
+        isFutureEffective: this.isFutureEffective
+      },
+      business: {
+        legalType: this.getEntityType,
+        identifier: this.getBusinessId
+      },
+      incorporationApplication: {
+        nameRequest: {
+          legalType: this.getEntityType
         },
-        business: {
-          legalType: this.getEntityType,
-          identifier: this.getBusinessId
+        nameTranslations: {
+          new: this.stateModel.nameTranslations
         },
-        incorporationApplication: {
-          nameRequest: {
-            legalType: this.getEntityType
-          },
-          nameTranslations: {
-            new: this.stateModel.nameTranslations
-          },
-          offices: this.stateModel.defineCompanyStep.officeAddresses,
-          contactPoint: {
-            email: this.stateModel.defineCompanyStep.businessContact.email,
-            phone: this.stateModel.defineCompanyStep.businessContact.phone,
-            extension: this.stateModel.defineCompanyStep.businessContact.extension
-          },
-          parties: this.getOrgPeople,
-          shareClasses: this.getShareClasses,
-          incorporationAgreement: {
-            agreementType: this.stateModel.incorporationAgreementStep.agreementType
-          }
+        offices: this.stateModel.defineCompanyStep.officeAddresses,
+        contactPoint: {
+          email: this.stateModel.defineCompanyStep.businessContact.email,
+          phone: this.stateModel.defineCompanyStep.businessContact.phone,
+          extension: this.stateModel.defineCompanyStep.businessContact.extension
+        },
+        parties: this.getOrgPeople,
+        shareStructure: {
+          shareClasses: this.getShareClasses
+        },
+        incorporationAgreement: {
+          agreementType: this.stateModel.incorporationAgreementStep.agreementType
         }
       }
     }
 
     // If this is a named IA then save Name Request Number and Approved Name.
     if (this.isNamedBusiness) {
-      filing.filing.incorporationApplication.nameRequest.nrNumber = this.getNameRequestNumber
-      filing.filing.incorporationApplication.nameRequest.legalName = this.getApprovedName
+      filing.incorporationApplication.nameRequest.nrNumber = this.getNameRequestNumber
+      filing.incorporationApplication.nameRequest.legalName = this.getApprovedName
     }
 
     // Pass the effective date only for a future effective filing.
@@ -97,27 +99,27 @@ export default class FilingTemplateMixin extends Vue {
     const effectiveDate: Date = this.getEffectiveDate
     if (effectiveDate) {
       const formattedDateTime = (effectiveDate.toISOString()).replace('Z', '+00:00')
-      filing.filing.header.effectiveDate = formattedDateTime
+      filing.header.effectiveDate = formattedDateTime
     }
     return filing
   }
 
   /**
-   * Parses an incorporation application filing into the store.
-   * @param filing the IA filing body to be parsed
+   * Parses an Incorporation Application filing into the store. Used when loading a filing.
+   * @param filing the IA filing body to parse
    */
-  parseIncorpApp (filing: IncorporationFilingBodyIF): void {
+  parseIncorpApp (filing: IncorporationFilingIF): void {
     // Set Business Information
     this.setBusinessInformation(filing.business)
 
     // Set Name Request
     this.setNameRequest(filing.incorporationApplication.nameRequest)
 
-    // Set Office Addresses
-    this.setOfficeAddresses(filing.incorporationApplication.offices)
-
     // Set Name Translations
     this.setNameTranslations(filing.incorporationApplication.nameTranslations?.new)
+
+    // Set Office Addresses
+    this.setOfficeAddresses(filing.incorporationApplication.offices)
 
     // Set Business Contact
     const contact = {
@@ -127,11 +129,20 @@ export default class FilingTemplateMixin extends Vue {
     this.setBusinessContact(contact)
 
     // Set Persons and Organizations
-    this.setOrgPersonList(filing.incorporationApplication.parties)
+    this.setOrgPersonList(filing.incorporationApplication.parties || [])
 
     // Set Share Structure
-    this.setShareClasses(filing.incorporationApplication.shareStructure
-      ? filing.incorporationApplication.shareStructure.shareClasses : [])
+    if (filing.incorporationApplication.shareStructure) {
+      this.setShareClasses(filing.incorporationApplication.shareStructure.shareClasses)
+    } else {
+      // if it exists, load data from old schema
+      const shareClasses = (filing.incorporationApplication as any).shareClasses
+      if (shareClasses) {
+        this.setShareClasses(shareClasses)
+      } else {
+        this.setShareClasses([])
+      }
+    }
 
     // Set Incorporation Agreement
     this.setIncorporationAgreementStepData({
@@ -181,5 +192,53 @@ export default class FilingTemplateMixin extends Vue {
   parseCorrection (filing: any): void {
     // See https://github.com/bcgov/business-schemas/blob/master/src/registry_schemas/schemas/correction.json
     // See https://github.com/bcgov/business-schemas/blob/master/src/registry_schemas/schemas/diff.json
+  }
+
+  /**
+    * Ensure consistent object structure for an incorporation application
+    * whether it contains a Name Request or not, and whether it is an initial
+    * draft or it has been previously saved. Object merging does not
+    * work very well otherwise (due to nested properties).
+    * @param filing the draft filing fetched from legal-api
+    * @returns the filing in safe-empty state if applicable
+  */
+  formatEmptyFiling (filing: any): IncorporationFilingIF {
+    let toReturn = filing
+    if (toReturn.incorporationApplication) {
+      // if there are no offices, populate empty array
+      if (!toReturn.incorporationApplication.offices) {
+        toReturn.incorporationApplication.offices = []
+      }
+
+      // if there is no contact point, populate empty object
+      if (!toReturn.incorporationApplication.contactPoint) {
+        toReturn.incorporationApplication.contactPoint = {
+          email: '',
+          phone: '',
+          extension: ''
+        }
+      }
+
+      // if there are no parties, populate empty array
+      if (!toReturn.incorporationApplication.parties) {
+        toReturn.incorporationApplication.parties = []
+      }
+
+      // if there is no share structure...
+      if (!toReturn.incorporationApplication.shareStructure) {
+        // if there are share classes (ie, old schema), assign them to the share structure (ie, new schema)
+        if (toReturn.incorporationApplication.shareClasses) {
+          toReturn.incorporationApplication.shareStructure = {
+            shareClasses: toReturn.incorporationApplication.shareClasses
+          }
+        } else {
+          // otherwise populate empty object
+          toReturn.incorporationApplication.shareStructure = {
+            shareClasses: []
+          }
+        }
+      }
+    }
+    return toReturn
   }
 }
